+++ HISTORY.rst
@@ -3,6 +3,15 @@
 History
 -------
 
+0.4.0 (2015-08-30)
+------------------
+
+* Added the following resources
+* Currency
+* Account Wallet
+* Account Dye Unlocks
+* Account Skin Unlocks
+
 0.3.0 (2015-07-06)
 ------------------